requirements.txt
fastapi
uvicorn
python-multipart
anthropic
pymupdf
python-dotenv
